Change SetMoveEffect into a function table and move to a new file
